Challan-Automation/requirements.txt
ast
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
bitarray==3.0.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
cytoolz==1.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
hexbytes==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parsimonious==0.10.0
pillow==11.0.0
propcache==0.2.0
psutil==6.1.0
psycopg2==2.9.10
pyclipper==1.3.0.post6
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
pywin32==308
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
scipy==1.14.1
setuptools==75.3.0
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.1
tifffile==2024.9.20
toolz==1.0.0
torch==2.5.1
tqdm==4.66.6
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
web3==7.4.0
websockets==13.1
Werkzeug==3.1.0
yarl==1.17.1
